requirements.txt
flask>=2.0.3
python-dotenv>=0.19.2
authlib>=1.0
requests>=2.27.1
flask-jwt-extended
datetime
timedelta
psycopg2-binary
werkzeug
flask_cors
flask_restx
networkx[default]